requirements.txt
pytest==7.1.1
PyYAML==6.0